reformat: take component out to folder, delete abandoned routes in app.tsx, services to services dedicated folder and centralization of types to types folder
