+++ frontend/src/App.tsx
@@ -1,10 +1,9 @@
 import { ThemeProvider, CssBaseline } from '@mui/material';
 import { lightTheme, darkTheme } from './theme';
 import { SharedProvider, useSharedContext } from './components/shared/context';
-import PriceListPage from './components/priceList/PriceListPage';
+import PriceListPage from './components/PriceListPage';
 import Layout from './components/shared/Layout';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import StationDetailPage from './components/stationDetail/StationDetailPage';
 
 function AppContent() {
   const { lightMode } = useSharedContext();
@@ -16,7 +15,6 @@ function AppContent() {
       <Layout title="FuelRank">
         <Routes>
           <Route path="/" element={<PriceListPage />} />
-          <Route path="/station/:id" element={<StationDetailPage />} />
         </Routes>
       </Layout>
     </ThemeProvider>
