--- frontend/src/components/stationDetail/StationDetailModal.tsx
@@ -1,191 +1,0 @@
-import { useEffect, useState } from 'react';
-import { useSharedContext } from '../shared/context';
-import { getStationDetail } from './service';
-import {
-  Avatar,
-  Box,
-  Card,
-  CardContent,
-  CardHeader,
-  Divider,
-  Modal,
-  Paper,
-  Typography,
-} from '@mui/material';
-import { cleanStationName } from '../../utils/stationNameCleaner';
-import StationMap from '../priceList/StationMap';
-
-type FuelPrice = {
-  price: number;
-  discount: number | null;
-  last_update: string;
-};
-
-export type StationDetail = {
-  id: string;
-  name: string;
-  brand: string;
-  address: string;
-  lat: number;
-  long: number;
-  url: string;
-  region: string;
-  prices: {
-    [fuelType: string]: FuelPrice;
-  };
-};
-
-const StationDetailModal = ({ stationId }: { stationId: string | null }) => {
-  const [station, setStation] = useState<StationDetail | null>(null);
-  const [error, setError] = useState<string | null>(null);
-  const { isMobile, modalState, setModalState } = useSharedContext();
-
-  const handleClose = () => setModalState(false);
-
-  useEffect(() => {
-    if (stationId) {
-      getStationDetail(stationId)
-        .then((res) => setStation(res))
-        .catch((err) => setError(err.message));
-    }
-  }, [stationId]);
-
-  const lastUpdate = station ? (Object.values(station.prices)[0]?.last_update ?? null) : null;
-
-  if (error) return <div>Error: {error}</div>;
-  //TODO: ERROR state
-  //TODO: LOADING state
-  if (!station)
-    return (
-      <Modal open={modalState} onClose={handleClose}>
-        <div>Loading...</div>
-      </Modal>
-    );
-
-  return (
-    <Modal open={modalState} onClose={handleClose}>
-      <Box
-        p={2}
-        maxWidth="md"
-        mx="auto"
-        sx={{
-          display: { xs: 'block', md: 'flex' },
-          flexDirection: 'column',
-          justifyContent: { md: 'center' },
-          minHeight: { xs: 'auto', md: '70vh' },
-          overflowY: { xs: 'visible', md: 'hidden' },
-        }}
-      >
-        <Card variant="outlined">
-          <CardHeader
-            sx={{ p: 1 }}
-            avatar={
-              <Avatar
-                src={`/logos/${station.brand.toLocaleLowerCase()}.png`}
-                variant="rounded"
-                sx={{ mb: 0.5, width: 56, height: 56 }}
-              />
-            }
-            title={
-              <Typography variant="h5">{cleanStationName(station.name, station.brand)}</Typography>
-            }
-            subheader={`${station.brand}, ${station.region}`}
-          />
-          <CardContent sx={{ pt: 1, pb: 0, '&:last-child': { pb: 0.25 } }}>
-            <Box
-              sx={{
-                display: 'flex',
-                flexDirection: { xs: 'column', md: 'row' },
-                gap: { xs: 0.5, md: 2 },
-                alignItems: 'stretch',
-              }}
-            >
-              {/*STATION INFO */}
-              <Box sx={{ flex: 1, display: 'flex', flexDirection: 'column' }}>
-                {isMobile && station && (
-                  <>
-                    <Divider sx={{ m: 1 }} />
-                    <Box sx={{ height: 170 }}>
-                      <StationMap
-                        selectedStation={{
-                          station_id: station.id,
-                          station_name: station.name,
-                          address: station.address,
-                          coords: [station.lat, station.long],
-                        }}
-                      />
-                    </Box>
-                    <Divider sx={{ my: 1 }} />
-                  </>
-                )}
-                <a
-                  href={`https://www.google.com/maps?q=${station.lat},${station.long}`}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  style={{ all: 'unset', cursor: 'pointer' }}
-                >
-                  <Box display="flex" alignItems="center" gap={1} my={1}>
-                    <img
-                      src="/map-icons/google-maps-icon.svg"
-                      alt="Open in Google maps"
-                      width={16}
-                      height={16}
-                      style={{ cursor: 'pointer' }}
-                    />
-                    <Typography>{station.address}</Typography>
-                  </Box>
-                </a>
-
-                {/**
-                  TODO: Links to official site from backend
-                {station.url && (
-                  <a
-                    href={station.url}
-                    target="_blank"
-                    rel="noopener"
-                    style={{ textDecoration: 'underline' }}
-                  >
-                    Visit official site
-                  </a>
-                )}
-                              */}
-              </Box>
-
-              {/* PRICES*/}
-              <Box
-                sx={{ flex: 1, display: 'flex', flexDirection: 'row', gap: 1, flexWrap: 'wrap' }}
-              >
-                {Object.entries(station.prices).map(([fuelType, fuel]) => (
-                  <Paper
-                    key={fuelType}
-                    sx={{
-                      flex: 1,
-                      p: { xs: 1, md: 2 },
-                      minWidth: { xs: '45%', md: '100%' },
-                      mb: { xs: 0, md: 2 },
-                    }}
-                    elevation={4}
-                  >
-                    <Typography variant="h6">{fuelType}</Typography>
-                    <Typography>Price: {fuel.price} ISK</Typography>
-
-                    {fuel.discount !== null && (
-                      <Typography>
-                        With discount: {(fuel.price - fuel.discount).toFixed(1)} ISK
-                      </Typography>
-                    )}
-                  </Paper>
-                ))}
-              </Box>
-              <Typography variant="caption" color="grey">
-                Last price update: {lastUpdate ? new Date(lastUpdate).toLocaleString() : 'N/A'}
-              </Typography>
-            </Box>
-          </CardContent>
-        </Card>
-      </Box>
-    </Modal>
-  );
-};
-
-export default StationDetailModal;
