+++ frontend/src/services/stationDetailService.ts
@@ -1,7 +1,6 @@
-import client from '../../api/client.ts';
+import client from '../api/client.ts';
 
-import type { StationDetail } from './StationDetailModal.tsx';
-
+import type { StationDetail } from '../types/types.ts';
 export const getStationDetail = (station_id: string): Promise<StationDetail> => {
   return client.get(`/stations/${station_id}`);
 };
